hexdump: size_t is better than uint64_t for cross-platform.

+++ src/hexdump.c
@@ -5,11 +5,11 @@
 #include <stdlib.h>
 #include <string.h>
 
-void hexdump(FILE *stream, const char *buf, const uint64_t len) {
+void hexdump(FILE *stream, const char *buf, const size_t len) {
 	if (!buf)
 		return;
 	int mod = 0;
-	uint64_t i = 0;
+	size_t i = 0;
 	char asciidata[17];
 	memset(asciidata, '.', 16);
 	asciidata[16] = 0;
@@ -41,7 +41,7 @@ void hexdump(FILE *stream, const char *buf, const uint64_t len) {
 }
 
 void named_hexdump(FILE *stream, const char *title,
-	const char *buf, const uint64_t len) {
+	const char *buf, const size_t len) {
 
 	fprintf(stream, "Hexdump: %s, %lu bytes\n", title, len);
 	hexdump(stream, buf, len);
